Mudado para Long o tipo do parâmetro de deletaUsuario()

+++ app/src/main/java/com/infact/nightour/controller/UsuariosController.java
@@ -92,7 +92,7 @@ public class UsuariosController {
         db.close();
     }
 
-    public void deletaUsuario(int id) {
+    public void deletaUsuario(Long id) {
         String where = Usuario.BD_ID + " = " + id;
 
         db = banco.getWritableDatabase();
